+++ service/controller/legacy/controller.go
@@ -41,6 +41,7 @@ type ControllerConfig struct {
 	PrometheusRemoteWriteURL    string
 	PrometheusRetentionDuration string
 	PrometheusRetentionSize     string
+	PrometheusScrapeInterval    string
 	PrometheusVersion           string
 
 	RestrictedAccessEnabled bool