+++ service/controller/managementcluster/controller.go
@@ -50,6 +50,7 @@ type ControllerConfig struct {
 	PrometheusLogLevel          string
 	PrometheusRetentionDuration string
 	PrometheusRetentionSize     string
+	PrometheusScrapeInterval    string
 	PrometheusVersion           string
 
 	RestrictedAccessEnabled bool