+++ service/controller/resource/monitoring/prometheus/resource_test.go
@@ -28,6 +28,7 @@ func TestPrometheus(t *testing.T) {
 		Registry:          "quay.io",
 		RetentionDuration: "2w",
 		RetentionSize:     "45Gi",
+		ScrapeInterval:    "60s",
 		Version:           "v2.28.1",
 	}