+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Remove Prometheus readinessProbe 5mn delay; since there is already a 15mn startupProbe
 
+### Changed
+
+- Increased ScrapeInterval from 30s to 60s
+
 ## [4.21.0] - 2023-02-23
 
 ### Changed